Consolidated code to single file, Added creds to gitignore, added pylint config
